[DisplayInfo] Introduced NETLINK socket connection
The current implementation does not update the hdmi properties
dynamically. This results in values being initialized once
and not updated on any hotplug events.

This commit introduces a ConnectionObserver, which will receive
any kernel messages to userspace - NETLINK_KOBJECT_UEVENT.
Upon receiving this message and determining that it relates
to a hotplug of HDMI, it will signal an event to a thread,
which in turn will re-query the properties using DRM C API.