[DisplayInfo] Refactored out part of C code.
The AMLogic implementation has been adjusted, so that
most of the C code and preprocessor macros are removed.
As this is the only implementation of it's own C functions,
there's no need for making them global and creating an artificial
C API.

The UpdateDisplay function has been reworked, in order to use
the first available mode from the HDMI connector.